requirements.txt
functions-framework==3.*
Pillow
numpy
tensorflow
flask-cors


